+++ CHANGELOG.md
@@ -7,8 +7,9 @@
 
 ### Added
 - registry/gc: Add Prometheus counter and histogram for online GC runs
-- registry/gc: Add Prometheus counter for online GC review postpones
+- registry/gc: Add Prometheus counter and histogram for online GC deletions
 - registry/gc: Add Prometheus counter for online GC deleted bytes
+- registry/gc: Add Prometheus counter for online GC review postpones
 
 ## [v3.3.0-gitlab] - 2021-04-09
 ### Added